修改mybatis插件

+++ src/main/java/com/yixun/ccmz/dao/mybatis/mapper/DictStdDiseaseMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface DictStdDiseaseMapper {
-    int countByExample(DictStdDiseaseExample example);
+    long countByExample(DictStdDiseaseExample example);
 
     int deleteByExample(DictStdDiseaseExample example);
